HEADER
-Added top nav bar.
-Styled header info

I was stuck on this section for over an hour and a half for some reason.
The nav bar was giving me trouble because the links were right next to each
other every time, but then I realized I can add gap to the flex styling.

Getting the right and left margins to line up also gave me trouble for the info
section. What I did was justify-content: center the div that contained the info and image,
then just added margins to the left and the right of the image and info. That seemed
to line up the left and right margins of the nav bar.